Added block for interpolated string transformation
Adds a block-capable stringForObject: that allows third-party developers
to transform an object's interpolated property values before injection
into the pattern.  This opens the door for adding percent escapes to
object value strings, without altering the overall validity of a url's
resource path.  Added documentation and test cases to showcase the new
functionality.  Additional cleansing is possible, with the assumption
that blocks are always available (iOS 4, newer OS X's), however in it's
current implementation it will function exactly as it did before
regardless.